+++ src/z80asm/options_def.h
@@ -76,8 +76,8 @@ OPT(OptStringList, &opts.lib_path, "-L", "--lib-path", "Add directory to library
 OPT(OptCallArg, option_define, "-D", "--define", "Define a static symbol", "SYMBOL")
 
 OPT_TITLE("Libraries:")
-OPT(OptCallOptArg, option_make_lib, "-x", "--make-lib", "Create a library file" FILEEXT_LIB, "[FILE]")
-OPT(OptCallOptArg, option_use_lib, "-i", "--use-lib", "Link library file" FILEEXT_LIB, "[FILE]")
+OPT(OptCallArg, option_make_lib, "-x", "--make-lib", "Create a library file" FILEEXT_LIB, "FILE")
+OPT(OptCallArg, option_use_lib, "-i", "--use-lib", "Link library file" FILEEXT_LIB, "FILE")
 
 OPT_TITLE("Binary Output:")
 OPT(OptSet, &opts.make_bin, "-b", "--make-bin", "Assemble and link/relocate to file" FILEEXT_BIN, "")